docs: Add test consolidation plan (Issue #185)
Amp-Thread-ID: https://ampcode.com/threads/T-019c48bb-ba28-71cb-98d5-e81f8fd61574
